% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{.wr_registry}
\alias{.wr_registry}
\title{Internal registry for retrievers}
\format{
An object of class \code{environment} of length 2.
}
\usage{
.wr_registry
}
\description{
Internal registry for retrievers
}
\keyword{internal}
