% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_client.R
\name{WebClient}
\alias{WebClient}
\title{WebClient: Minimal client for DuckDuckGo + LLM chat}
\description{
Wraps DuckDuckGo Instant Answer search and a user-supplied chat client
(e.g., \code{ellmer::chat_aws_bedrock()}), then exposes \code{$ask()}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WebClient-new}{\code{WebClient$new()}}
\item \href{#method-WebClient-set_chat}{\code{WebClient$set_chat()}}
\item \href{#method-WebClient-search}{\code{WebClient$search()}}
\item \href{#method-WebClient-context}{\code{WebClient$context()}}
\item \href{#method-WebClient-ask}{\code{WebClient$ask()}}
\item \href{#method-WebClient-clone}{\code{WebClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-WebClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebClient$new(
  k_web = getOption("webretrieveR.k_web", 5),
  chat_client = getOption("webretrieveR.chat_client"),
  system_rules = getOption("webretrieveR.system_rules")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebClient-set_chat"></a>}}
\if{latex}{\out{\hypertarget{method-WebClient-set_chat}{}}}
\subsection{Method \code{set_chat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebClient$set_chat(chat_client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebClient-search"></a>}}
\if{latex}{\out{\hypertarget{method-WebClient-search}{}}}
\subsection{Method \code{search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebClient$search(question)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebClient-context"></a>}}
\if{latex}{\out{\hypertarget{method-WebClient-context}{}}}
\subsection{Method \code{context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebClient$context(question, hits = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebClient-ask"></a>}}
\if{latex}{\out{\hypertarget{method-WebClient-ask}{}}}
\subsection{Method \code{ask()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebClient$ask(question)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WebClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
