% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wr_build_context}
\alias{wr_build_context}
\title{Build a compact context block from bullets}
\usage{
wr_build_context(source, query, bullets, max_chars = 1500)
}
\description{
Build a compact context block from bullets
}
\keyword{internal}
