% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_ddg.R
\name{wr_engine_ddg}
\alias{wr_engine_ddg}
\title{DuckDuckGo Instant Answer retriever}
\usage{
wr_engine_ddg(
  q,
  max_chars = 1500,
  include_urls = TRUE,
  app_id = "webretrieveR/0.1",
  timeout_sec = 6
)
}
\description{
DuckDuckGo Instant Answer retriever
}
\keyword{internal}
