% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{wr_register_engine}
\alias{wr_register_engine}
\title{Register a retriever engine}
\usage{
wr_register_engine(name, fun)
}
\arguments{
\item{name}{Engine name (e.g., "ddg", "wikipedia").}

\item{fun}{Function implementing the retriever signature.}
}
\description{
A retriever is a function of the form:
function(q, max_chars = 1500, include_urls = TRUE, ...) -> character
returning a single compact context string.
}
