% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_wikipedia.R
\name{wr_engine_wikipedia}
\alias{wr_engine_wikipedia}
\title{Wikipedia retriever (search top hit -> summary)}
\usage{
wr_engine_wikipedia(
  q,
  max_chars = 1500,
  include_urls = TRUE,
  lang = "en",
  timeout_sec = 6
)
}
\description{
Wikipedia retriever (search top hit -> summary)
}
\keyword{internal}
